Fluent validation added custom Exception handler added

+++ HR.LeaveManagenet.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,7 @@
-﻿using AutoMapper;
-using HR.LeaveManagment.Application.Features.LeaveAllocations.Requests.Commands;
-using HR.LeaveManagment.Application.Persistence.Contracts;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
+using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -8,7 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HR.LeaveManagment.Application.Features.LeaveAllocations.Handlers.Commands
+namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands
 {
     public class DeleteLeaveAllocationCommandHandler: IRequestHandler<DeleteLeaveAllocationCommand>
     {
@@ -21,6 +22,10 @@ namespace HR.LeaveManagment.Application.Features.LeaveAllocations.Handlers.Comma
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+            }
             await _leaveAllocationRepository.Delete(leaveAllocation);
             return Unit.Value;
         }
